+++ src/argilla/client/sdk/v1/workspaces/api.py
@@ -12,7 +12,7 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from typing import Union
+from typing import List, Union
 from uuid import UUID
 
 import httpx
@@ -64,6 +64,34 @@ def delete_workspace(
 
     if response.status_code == 200:
         parsed_response = WorkspaceModel(**response.json())
+        return Response(
+            status_code=response.status_code,
+            content=response.content,
+            headers=response.headers,
+            parsed=parsed_response,
+        )
+    return handle_response_error(response)
+
+
+def list_workspaces_me(
+    client: httpx.Client,
+) -> Response[Union[List[WorkspaceModel], ErrorMessage, HTTPValidationError]]:
+    """Sends a GET request to `/api/v1/me/workspaces` endpoint to get the list of
+    workspaces the current user has access to.
+
+    Args:
+        client: the authenticated Argilla client to be used to send the request to the API.
+
+    Returns:
+        A `Response` object containing a `parsed` attribute with the parsed response if
+        the request was successful, which is a list of `WorkspaceModel`.
+    """
+    url = "/api/v1/me/workspaces"
+
+    response = client.get(url=url)
+
+    if response.status_code == 200:
+        parsed_response = [WorkspaceModel(**workspace) for workspace in response.json()["items"]]
         return Response(
             status_code=response.status_code,
             content=response.content,